Improves endpoint error handling

+++ eth_crowdfund_api/src/views/CampaignView.py
@@ -1,5 +1,7 @@
 from flask import request, json, Response, Blueprint, jsonify
 from ..models.Campaign import Campaign, CampaignSchema
+from marshmallow import ValidationError
+from sqlalchemy import exc
 from . import custom_response
 
 
@@ -8,13 +10,13 @@ campaign_schema = CampaignSchema()
 
 @campaign_api.route('/', methods=['POST'])
 def create():
-  req_data = request.get_json()
-  data = campaign_schema.load(req_data)
-  # data, error = campaign_schema.load(req_data)
-  # BUT WITH ERROR: the above code threw an error, currently ignoring errors
-  error = None
-  if error: 
-    return custom_response(error, 400)
+  try: 
+    req_data = request.get_json()
+    data = campaign_schema.load(req_data)
+  except ValidationError as err: 
+    return custom_response(err.messages, 400)
+  except: 
+    return custom_response({ "error": "No input data provided" }, 400)
 
   campaign = Campaign(data)
   campaign.save() 
@@ -24,7 +26,7 @@ def create():
 @campaign_api.route('/<int:campaign_id>', methods=['GET'])
 def get_a_campaign(campaign_id):
   campaign = Campaign.get_one_campaign(campaign_id)
-  if not campaign: 
+  if campaign is None: 
     return custom_response({'error':'Campaign not found'}, 404)
 
   campaign_data = campaign_schema.dump(campaign)
@@ -32,15 +34,18 @@ def get_a_campaign(campaign_id):
 
 @campaign_api.route('/<int:campaign_id>', methods=['PUT'])
 def update(campaign_id):
-  req_data = request.get_json()
-  data = campaign_schema.load(req_data, partial=True)
-  # error not functional yet
-  error = None
-  if error:
-    return custom_response(error, 400)
-  
-  campaign = Campaign.get_one_campaign(campaign_id)
-  campaign.update(data)
+  try: 
+    req_data = request.get_json()
+    data = campaign_schema.load(req_data, partial=True)
+    campaign = Campaign.get_one_campaign(campaign_id)
+    campaign.update(data)
+  except ValidationError as err: 
+    return custom_response(err.messages, 400)
+  except exc.IntegrityError as err: 
+    return custom_response({ "error": err.orig.diag.message_detail}, 400)
+  except: 
+    return custom_response({ "error": "No input data provided" }, 400)
+
   campaign_data = campaign_schema.dump(campaign)
   return custom_response(campaign_data, 200)
 
@@ -48,6 +53,10 @@ def update(campaign_id):
 @campaign_api.route('/<int:campaign_id>', methods=["DELETE"])
 def delete(campaign_id):
   campaign = Campaign.get_one_campaign(campaign_id)
+
+  if campaign is None: 
+    return custom_response({ 'error': 'Campaign not found'}, 404)
+
   campaign.delete()
   campaign_data = campaign_schema.dump(campaign)
   return custom_response(campaign_data, 200)
@@ -56,10 +65,9 @@ def delete(campaign_id):
 def get_all_campaigns():
   campaigns = Campaign.get_all_campaigns()
 
-  if not campaigns:
-    return custom_response({'error': 'No Campaigns'}, 404)
+  if campaigns is None:
+    return custom_response({'error': 'No Campaigns found'}, 404)
 
-  # campaign_data = campaign_schema.dump(campaigns)
   campaign_data = []
   for campaign in campaigns:
     campaign_data.append(campaign_schema.dump(campaign))
