+++ eth_crowdfund_api/src/views/RequestView.py
@@ -1,5 +1,6 @@
 from flask import request, json, Response, Blueprint, g, jsonify
 from marshmallow import ValidationError
+from sqlalchemy import exc
 from ..models.Request import Request, RequestSchema
 from . import custom_response
 
@@ -26,8 +27,8 @@ def create():
 def get_all_requests():
   requests = Request.get_all_requests()
 
-  if not requests:
-    return custom_response({'error': 'No Requests'}, 404)
+  if requests is None:
+    return custom_response({'error': 'No Requests found'}, 404)
 
   request_data = []
   for campaign_request in requests:
@@ -37,7 +38,7 @@ def get_all_requests():
 @request_api.route('/<int:request_id>', methods=['GET'])
 def get_a_request(request_id):
   campaign_request = Request.get_one_request(request_id)
-  if not campaign_request:
+  if campaign_request is None:
     return custom_response({ 'error': 'Request not found'}, 404)
   
   campaign_request_data = request_schema.dump(campaign_request)
@@ -45,17 +46,29 @@ def get_a_request(request_id):
 
 @request_api.route('/<int:request_id>', methods=['PUT'])
 def update(request_id):
-  req_data = request.get_json()
-  data = request_schema.load(req_data, partial=True)
 
-  campaign_request = Request.get_one_request(request_id)
-  campaign_request.update(data)
+  try: 
+    req_data = request.get_json()
+    data = request_schema.load(req_data, partial=True)
+    campaign_request = Request.get_one_request(request_id)
+    campaign_request.update(data)
+  except ValidationError as err:
+    return custom_response(err.messages, 400)
+  except exc.IntegrityError as err: 
+    return custom_response({ "error": err.orig.diag.message_detail }, 400)
+  except:  
+    return custom_response({ "error": "No input data provided" }, 400)
+
   request_data = request_schema.dump(campaign_request)
   return custom_response(request_data, 200)
 
 @request_api.route('/<int:request_id>', methods=['DELETE'])
 def delete(request_id):
   campaign_request = Request.get_one_request(request_id)
+
+  if campaign_request is None: 
+    return custom_response({ 'error': 'Request not found'}, 404)
+
   campaign_request.delete()
   campaign_request_data = request_schema.dump(campaign_request)
   return custom_response(campaign_request_data, 200)
